Moved EffectAllowed to shared so that it could be used in shared state directly
